Qtile: Implement floating `neorg` notes
Differentiate `neorg` windows which require different floating rules by
`WM_NAME` and the absence or presence of a defined `WM_WINDOW_ROLE`.

The `Match` object, that is used when setting the `float_rules`, returns
`true` even on substring matches. I.e. It can find a match even when the
window client attribute is not exactly the same as the `Match` object's
attribute.